API

Public

COVID_Tracking_Clean.COVID_Tracking_Clean — Module

COVID_Tracking_Clean

This module is meant to clean the data from the COVID Tracking Project.

COVID_Tracking_Clean.main — Function

main()

Performs the action to refresh the data. Latest data will be written to data/latest.tsv

Private

COVID_Tracking_Clean.GITHUB_API_QUERY — Constant

GITHUB_API_QUERY

GitHub GraphQL query.

COVID_Tracking_Clean.GITHUB_GRAPHQL_ENDPOINT — Constant

GITHUB_GRAPHQL_ENDPOINT::String = "https://api.github.com/graphql"

GitHub API v4 GraphQL API endpoint.

COVID_Tracking_Clean.GITHUB_REST_ENDPOINT — Constant

GITHUB_REST_ENDPOINT::String = "https://api.github.com"

GitHub API v3 RESTful root endpoint.

COVID_Tracking_Clean.API_Limits — Type

API_Limits

GitHub API limits for a PersonalAccessToken.

Fields

core::Limits

search::Limits

graphql::Limits

COVID_Tracking_Clean.GitHubPersonalAccessToken — Type

GitHubPersonalAccessToken(login::AbstractString,
                          token::AbstractString
                          )::GitHubPersonalAccessToken

A GitHub Personal Access Token

Fields

login::String

token::String

client::Client

limits::Limits

COVID_Tracking_Clean.Limits — Type

Limits

GitHub API limits.

It includes how many remaining queries are available for the current time period and when it resets.

Fields

limit::Int

remaining::Int

reset::ZonedDateTime

COVID_Tracking_Clean.Opt — Type

Opt(login::AbstractString,
    token::AbstractString;
    db_usr::AbstractString = "postgres",
    db_pwd::AbstractString = "postgres",
    host::AbstractString = "postgres",
    port::Integer = 5432,
    dbname::AbstractString = "postgres",
    schema::AbstractString = "postgres",
    role::AbstractString = "postgres"
    )::Opt

Structure for passing arguments to functions.

Fields

conn::Connection

schema::String

role::String

pat::GitHubPersonalAccessToken

Example

julia> opt = Opt("Nosferican",
                 ENV["GITHUB_TOKEN"],
                 host = ENV["POSTGRES_HOST"],
                 port = parse(Int, ENV["POSTGRES_PORT"]));

COVID_Tracking_Clean.find_shas — Function

find_shas(id::AbstractString = "MDEwOlJlcG9zaXRvcnkyNDY0MTE2MDc=",
          directory::AbstractString = "data",
          file::AbstractString = "states_current.csv",
          just_last::Bool = false)

Return slug for the repository and SHA1 for the file at directory/file (all versions).

COVID_Tracking_Clean.get_sha — Function

get_sha(obj, directory::AbstractString = "data", file::AbstractString = "states_current.csv")::String

Navigate the entries in the tree to find the SHA1 of the file at path (directory/file)

COVID_Tracking_Clean.get_tbl — Method

get_tbl(sha::AbstractString)

Uses the SHA1 to download from GitHub the version of COVID19Tracking/covid-tracking-data/data/states_current.csv

COVID_Tracking_Clean.graphql — Method

graphql(obj::GitHubPersonalAccessToken,
        operationName::AbstractString,
        vars::Dict{String})

Return JSON of the GraphQL query.

COVID_Tracking_Clean.prune — Method

prune(tbl)::DataFrame

Simplifies the grade tsrange.

COVID_Tracking_Clean.restful — Method

restful(obj::GitHubPersonalAccessToken,
        endpoint::AbstractString;
        method::AbstractString = "GET",
        params::Dict{String} = Dict{String,String}(),
)

Return response.

COVID_Tracking_Clean.states_daily — Method

states_daily()

Uses the Covid Tracking API to return the states/daily table.